+++ trunk/server/TPFC_device_3dstereo.h
@@ -19,19 +19,19 @@ class TPFC_device_3dstereo : public TPFC_device{
     int getsourcepos(TPFC_device* s);
     // funcion que convierte el angulo del device fuente apuntado por sources[2o parametro]
     // a un angulo usable para coordenar (el angulo interior del triangulo usado)
-    float angleconversion(float, int);
+    double angleconversion(double, int);
     
 
     // variables usadas en la calibracion
     bool calibrated; // flag de calibrado
-    float* calib_data; // buffer que se usa durante el calibrado. los datos estan almacenados
+    double* calib_data; // buffer que se usa durante el calibrado. los datos estan almacenados
 		       // en tramos de calib_sample longitud, en este orden:
 		       // x_wm1_dot1, y_wm1_dot1, x_wm2_dot1, y_wm2_dot1,
 		       // x_wm1_dot2, y_wm1_dot2, x_wm2_dot2, y_wm2_dot2... 
     int calib_count; // total de puntos almacenados en calib_data (indice)
     pthread_mutex_t* calib_lock;
-    void addsample(float*); // funcion para añadir los datos de una muestra al bufer
-    float* getsamples(int, int, int); // funcion para recuperar todos los samples del buffer
+    void addsample(double*); // funcion para añadir los datos de una muestra al bufer
+    double* getsamples(int, int, int); // funcion para recuperar todos los samples del buffer
 
     public:
     // consctructora y creadora
